bug fixes from simulator testing
Note that the current values do not work very well in the simulator, they were
tuned for the real world. The logic could also be improved, this is just a
direct conversion of the logic we had at the pool test but now using helper
classes and is better organized.